+++ api/src/main/java/energy/eddie/api/CommonInformationModelVersions.java
@@ -4,7 +4,7 @@ package energy.eddie.api;
  * The supported versions of the CIM
  */
 public enum CommonInformationModelVersions {
-    V0_82("0.82"), V0_91_08("0.91.08"), V0_92_08("0.92.08");
+    V0_82("0.82"), V0_91_08("0.91.08"), V1_04("1.04");
 
     private final String version;