+++ api/src/main/java/energy/eddie/api/v1_04/outbound/ValidatedHistoricalDataMarketDocumentOutboundConnector.java
@@ -1,6 +1,6 @@
-package energy.eddie.api.v0_92_08.outbound;
+package energy.eddie.api.v1_04.outbound;
 
-import energy.eddie.cim.v0_92_08.VHDEnvelope;
+import energy.eddie.cim.v1_04.vhd.VHDEnvelope;
 import reactor.core.publisher.Flux;
 
 /**